requirements.txt
flask===2.3.2
flask_sqlalchemy===3.1.1
python-dotenv===1.0.0
psycopg2-binary===2.9.9